Read description please
This commit contains functionality to allow the player to be in a state where only camera movement is allowed, as per design request.

I just wanted to point out that for some reason this commit contains a few .meta changes, some of which concern objects/scenes that I never touched. Should be completely inconsequential, just wanted to make sure