Refactored note system, details in description
Note system is now working in a way that's much closer to item inventory.
This means notes don't need to be all declared and defined between the manager class anymore, and that the entry list will be empty if no notes are in.
A user can specify what note a pickup contains by populating the NoteInGameObjectScript.
Further functions can be added or brought back, not really needed for now however.

Also fixed a couple redundant calls and trimmed code that was difficult to read.